requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
PyJWT==2.8.0
qrcode[pil]==7.4.2
Pillow==10.1.0
cryptography>=41.0.0
aiohttp>=3.8.0

